// apps/web/src/shared/api/graphql/__generated__/schema.graphql.ts
export type Maybe<T> = T | null;
// ...
export interface Scalars {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Date: { input: any; output: any };
  DateTime: { input: any; output: any };
  JSON: { input: any; output: any };
  Time: { input: any; output: any };
  Upload: { input: any; output: any };
}
// ...
export interface Pagination {
  __typename?: 'Pagination';
  page: Scalars['Int']['output'];
  pageCount: Scalars['Int']['output'];
  pageSize: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
}
// ...
export interface ResponseCollectionMeta {
  __typename?: 'ResponseCollectionMeta';
  pagination: Pagination;
}

export interface Schedule {
  __typename?: 'Schedule';
  artist?: Maybe<Artist>;
  time?: Maybe<Scalars['String']['output']>;
}

export interface ShopCategory {
  __typename?: 'ShopCategory';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  parent?: Maybe<ShopCategoryRelationResponseCollection>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}

export interface ShopCategoryParentArgs {
  filters?: InputMaybe<ShopCategoryFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
}

export interface ShopCategoryEntity {
  __typename?: 'ShopCategoryEntity';
  attributes?: Maybe<ShopCategory>;
  id?: Maybe<Scalars['ID']['output']>;
}
// ...
export interface ShopCategoryRelationResponseCollection {
  __typename?: 'ShopCategoryRelationResponseCollection';
  data: Array<ShopCategoryEntity>;
}

export interface ShopItem {
  __typename?: 'ShopItem';
  attributes?: Maybe<Scalars['JSON']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  images: UploadFileRelationResponseCollection;
  oldPrice?: Maybe<Scalars['Int']['output']>;
  price: Scalars['Int']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  quantity?: Maybe<Scalars['Int']['output']>;
  shop_category?: Maybe<ShopCategoryEntityResponse>;
  slug?: Maybe<Scalars['String']['output']>;
  soldout?: Maybe<Scalars['Boolean']['output']>;
  title: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}

export interface ShopItemImagesArgs {
  filters?: InputMaybe<UploadFileFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
}

export interface ShopItemEntity {
  __typename?: 'ShopItemEntity';
  attributes?: Maybe<ShopItem>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface ShopItemEntityResponse {
  __typename?: 'ShopItemEntityResponse';
  data?: Maybe<ShopItemEntity>;
}

export interface ShopItemEntityResponseCollection {
  __typename?: 'ShopItemEntityResponseCollection';
  data: Array<ShopItemEntity>;
  meta: ResponseCollectionMeta;
}